+++ protocols/src/main/scala/dx/util/protocols/DxFileAccessProtocol.scala
@@ -1,17 +1,23 @@
 package dx.util.protocols
 
 import java.nio.charset.Charset
-import java.nio.file.Path
-
-import dx.api.{DxApi, DxFile, DxFileDescCache}
-import dx.util.{AbstractAddressableFileNode, FileAccessProtocol, FileSource, FileUtils}
+import java.nio.file.{Files, Path, Paths}
+import dx.api.{DxApi, DxFile, DxFileDescCache, DxFindDataObjects, DxPath, DxProject}
+import dx.util.{
+  AbstractAddressableFileNode,
+  AddressableFileSource,
+  FileAccessProtocol,
+  FileSource,
+  FileUtils,
+  SysUtils
+}
 
 case class DxFileSource(dxFile: DxFile, override val encoding: Charset)(
     override val address: String,
     dxApi: DxApi
 ) extends AbstractAddressableFileNode(address, encoding) {
 
-  override def name: String = dxFile.describe().name
+  override def name: String = dxFile.getName
 
   override def folder: String = dxFile.describe().folder
 
@@ -23,6 +29,58 @@ case class DxFileSource(dxFile: DxFile, override val encoding: Charset)(
 
   override protected def localizeTo(file: Path): Unit = {
     dxApi.downloadFile(file, dxFile)
+  }
+}
+
+case class DxArchiveFolderSource(dxFileSource: DxFileSource) extends AddressableFileSource {
+  override def isDirectory: Boolean = true
+
+  override def address: String = dxFileSource.address
+
+  override def name: String = {
+    FileUtils.changeFirstFileExt(dxFileSource.name, Vector(".tar.gz", ".tgz", ".tar"))
+  }
+
+  override def folder: String = dxFileSource.folder
+
+  override protected def localizeTo(dir: Path): Unit = {
+    val tempfile = Files.createTempFile("temp", dxFileSource.name)
+    val tarOpts = if (dxFileSource.name.endsWith("gz")) "-xz" else "-x"
+    try {
+      dxFileSource.localize(tempfile)
+      SysUtils.execCommand(s"tar ${tarOpts} -f ${tempfile} -C ${dir.toString} --strip-components=1")
+    } finally {
+      tempfile.toFile.delete()
+    }
+  }
+}
+
+case class DxFolderSource(dxProject: DxProject, folder: String)(
+    dxApi: DxApi
+) extends AddressableFileSource {
+  override def address: String = s"dx://${dxProject.id}:${folder}"
+
+  override def name: String = folder
+
+  override def isDirectory: Boolean = true
+
+  lazy val listing: Vector[(DxFile, Path)] = {
+    val results =
+      DxFindDataObjects(dxApi).apply(Some(dxProject), Some(folder), recurse = true, Some("file"))
+    results.map {
+      case (f: DxFile, _) =>
+        val relPath = Paths.get(folder).relativize(Paths.get(f.describe().folder))
+        (f, relPath)
+      case other => throw new Exception(s"unexpected result ${other}")
+    }.toVector
+  }
+
+  override protected def localizeTo(dir: Path): Unit = {
+    listing.foreach {
+      case (dxFile, relPath) =>
+        val path = dir.resolve(relPath).resolve(dxFile.getName)
+        dxApi.downloadFile(path, dxFile)
+    }
   }
 }
 
@@ -38,6 +96,8 @@ case class DxFileAccessProtocol(dxApi: DxApi = DxApi.get,
     extends FileAccessProtocol {
   override val schemes = Vector(DxFileAccessProtocol.DxUriScheme)
   private var uriToFileSource: Map[String, DxFileSource] = Map.empty
+
+  override val supportsDirectories: Boolean = true
 
   private def resolveFileUri(uri: String): DxFile = {
     uri.split("::").toVector match {
@@ -60,15 +120,31 @@ case class DxFileAccessProtocol(dxApi: DxApi = DxApi.get,
     }
   }
 
+  private def resolveFile(uri: String): DxFileSource = {
+    val dxFile = dxFileCache.updateFileFromCache(resolveFileUri(uri))
+    val src = DxFileSource(dxFile, encoding)(uri, dxApi)
+    uriToFileSource += (uri -> src)
+    src
+  }
+
   override def resolve(uri: String): DxFileSource = {
     // First search in the cache. This may save us an API call.
     uriToFileSource.get(uri) match {
       case Some(src) => src
-      case None =>
-        val dxFile = dxFileCache.updateFileFromCache(resolveFileUri(uri))
-        val src = DxFileSource(dxFile, encoding)(uri, dxApi)
-        uriToFileSource += (uri -> src)
-        src
+      case None      => resolveFile(uri)
+    }
+  }
+
+  override def resolveDirectory(uri: String): AddressableFileSource = {
+    // a Directory may be either a dx file or a dx://project:/path/to/dir/ URI.
+    if (uri.endsWith("/")) {
+      val (projectName, folder) = DxPath.split(uri)
+      val project = projectName
+        .map(dxApi.resolveProject)
+        .getOrElse(throw new Exception("project must be specified for a DNAnexus folder URI"))
+      DxFolderSource(project, folder)(dxApi)
+    } else {
+      DxArchiveFolderSource(resolveFile(uri))
     }
   }